Create App folder
