+++ App/Controllers/AttendingEventController.php
@@ -1,31 +1,28 @@
 <?php
 
-namespace Controllers;
+namespace App\Controllers;
 
-/* TODO: Test controller */
-
-class TeamController {
-    public function index ($request, $response, $args)
+class AttendingEventController {
+    public function updateSubscription ($request, $response, $args)
     {
-
         authAllowAuthenticated();
 
         $aEventId 	= isset($_REQUEST['event']) ? (int)$_REQUEST['event'] : '';
         $aAction 	= isset($_REQUEST['action']) ? $_REQUEST['action'] : '';
         $aRet 		= '';
-
+        
         header('Content-Type: text/html; charset=utf-8');
-
+        
         switch($aAction) {
             case 'subscribe':
                 $aEvent = eventGetById($aEventId);
                 $aUser	= authGetAuthenticatedUserInfo();
-
+        
                 if ($aEvent != null) {
                     try {
                         attendingAdd($aUser['id'], $aEvent['id'], 0);
-                        $aRet = '<span class="label label-success">Inscrito</span>';
-                        $aRet .= ' <a href="#" onclick="SAC.unsubscribe('.$aEventId.')">[X]</a>';
+                        $aRet = '<span class="label label-success"><i class="fa fa-check-square"></i> Inscrito</span>';
+                        $aRet .= ' <a href="javascript:void(0);" onclick="SAC.unsubscribe('.$aEventId.')" title="Clique para remover sua inscrição dessa atividade."><i class="fa fa-remove"></i></a>';
                         
                     } catch(Exception $aError) {
                         $aRet = 'Oops! ' . $aError->getMessage();
@@ -39,11 +36,11 @@ class TeamController {
             case 'unsubscribe':
                 $aEvent = eventGetById($aEventId);
                 $aUser	= authGetAuthenticatedUserInfo();
-
+        
                 if ($aEvent != null) {
                     try {
                         attendingRemove($aUser['id'], $aEvent['id'], 0);
-                        $aRet = '<a href="#" onclick="SAC.subscribe('.$aEventId.', '.($aEvent['capacity'] != 0 ? 'true' : 'false').')">[S]</a>';
+                        $aRet = '<a href="javascript:void(0);" onclick="SAC.subscribe('.$aEventId.', '.($aEvent['capacity'] != 0 ? 'true' : 'false').')" title="Clique para se inscrever nessa atividade."><i class="fa fa-square-o"></i></a>';
                         
                     } catch(Exception $aError) {
                         $aRet = 'Oops! ' . $aError->getMessage();
@@ -57,7 +54,8 @@ class TeamController {
             default:
                 $aRet['msg'] = 'Opção inválida!';
         }
-
+        
         echo $aRet;
+        return $response;
     }
 }
