+++ App/Controllers/EventController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Controllers;
+namespace App\Controllers;
 
 use \core\View;
