+++ App/Controllers/HomeController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Controllers;
+namespace App\Controllers;
 
 class HomeController {
     public function home ($request, $response, $args)
