+++ public/index.php
@@ -12,25 +12,25 @@
 
 	$app->setBasePath('/sac');
 
-	$app->get('/', 'Controllers\HomeController:home');
+	$app->get('/', 'App\Controllers\HomeController:home');
 
-	$app->get('/api/payment/', 'Controllers\PaymentController:apiIndex');
-	$app->post('/api/attending-event/update-subscription', 'Controllers\AttendingEventController:updateSubscription');
+	$app->get('/api/payment/', 'App\Controllers\PaymentController:apiIndex');
+	$app->post('/api/attending-event/update-subscription', 'App\Controllers\AttendingEventController:updateSubscription');
 
-	$app->get('/login', 'Controllers\AuthController:loginForm');
-	$app->get('/inscricao', 'Controllers\AuthController:subscriptionForm');
-	$app->post('/login', 'Controllers\AuthController:login');
-	$app->get('/logout', 'Controllers\AuthController:logout');
+	$app->get('/login', 'App\Controllers\AuthController:loginForm');
+	$app->get('/inscricao', 'App\Controllers\AuthController:subscriptionForm');
+	$app->post('/login', 'App\Controllers\AuthController:login');
+	$app->get('/logout', 'App\Controllers\AuthController:logout');
 
-	$app->get('/admin/evento', 'Controllers\EventController:adminIndex');
-	$app->post('/admin/evento', 'Controllers\EventController:create');
-	$app->get('/admin/campeonato', 'Controllers\CompetitionController:index');
-	$app->post('/admin/campeonato', 'Controllers\CompetitionController:create');
-	$app->get('/admin/inscricoes', 'Controllers\SubscribeController:index');
-	$app->get('/admin/pagamento', 'Controllers\SubscribeController:payment');
-	$app->post('/admin/pagamento', 'Controllers\SubscribeController:paymentCreate');
+	$app->get('/admin/evento', 'App\Controllers\EventController:adminIndex');
+	$app->post('/admin/evento', 'App\Controllers\EventController:create');
+	$app->get('/admin/campeonato', 'App\Controllers\CompetitionController:index');
+	$app->post('/admin/campeonato', 'App\Controllers\CompetitionController:create');
+	$app->get('/admin/inscricoes', 'App\Controllers\SubscribeController:index');
+	$app->get('/admin/pagamento', 'App\Controllers\SubscribeController:payment');
+	$app->post('/admin/pagamento', 'App\Controllers\SubscribeController:paymentCreate');
 
-	$app->get('/times', 'Controllers\TeamController:index');
+	$app->get('/times', 'App\Controllers\TeamController:index');
 
 	// Run app
 	$app->run();
